+++ src/com/intellij/idea/plugin/hybris/polyglotQuery/PolyglotQueryLanguage.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -21,10 +21,10 @@ package com.intellij.idea.plugin.hybris.polyglotQuery
 import com.intellij.lang.Language
 import java.io.Serial
 
-class PolyglotQueryLanguage : Language("Polyglot Query") {
-    companion object {
-        @Serial
-        private val serialVersionUID: Long = -8539604107669994901L
-        val instance = PolyglotQueryLanguage()
-    }
+object PolyglotQueryLanguage : Language("Polyglot Query") {
+
+    private fun readResolve(): Any = PolyglotQueryLanguage
+
+    @Serial
+    private val serialVersionUID: Long = -8539604107669994901L
 }