+++ src/com/intellij/idea/plugin/hybris/polyglotQuery/file/PolyglotQueryFile.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -23,7 +23,7 @@ import com.intellij.idea.plugin.hybris.polyglotQuery.PolyglotQueryLanguage
 import com.intellij.psi.FileViewProvider
 import java.io.Serial
 
-class PolyglotQueryFile(viewProvider: FileViewProvider) : PsiFileBase(viewProvider, PolyglotQueryLanguage.instance) {
+class PolyglotQueryFile(viewProvider: FileViewProvider) : PsiFileBase(viewProvider, PolyglotQueryLanguage) {
 
     override fun getFileType() = PolyglotQueryFileType