+++ src/com/intellij/idea/plugin/hybris/polyglotQuery/formatting/PgQSpacingBuilder.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -24,7 +24,7 @@ import com.intellij.idea.plugin.hybris.polyglotQuery.psi.PolyglotQueryTypes.*
 import com.intellij.psi.codeStyle.CodeStyleSettings
 import com.intellij.psi.tree.TokenSet
 
-class PgQSpacingBuilder(settings: CodeStyleSettings) : SpacingBuilder(settings, PolyglotQueryLanguage.instance) {
+class PgQSpacingBuilder(settings: CodeStyleSettings) : SpacingBuilder(settings, PolyglotQueryLanguage) {
 
     init {
         this