+++ src/com/intellij/idea/plugin/hybris/system/cockpitng/model/config/hybris/Labels.java
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -22,11 +22,7 @@
 package com.intellij.idea.plugin.hybris.system.cockpitng.model.config.hybris;
 
 import com.intellij.idea.plugin.hybris.common.HybrisConstants;
-import com.intellij.util.xml.DomElement;
-import com.intellij.util.xml.GenericAttributeValue;
-import com.intellij.util.xml.GenericDomValue;
-import com.intellij.util.xml.Namespace;
-import com.intellij.util.xml.SubTag;
+import com.intellij.util.xml.*;
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -34,6 +30,8 @@ import org.jetbrains.annotations.NotNull;
  */
 @Namespace(HybrisConstants.COCKPIT_NG_NAMESPACE_KEY)
 public interface Labels extends DomElement {
+
+    String LABEL = "label";
 
     /**
      * Returns the value of the beanId child.
@@ -51,7 +49,7 @@ public interface Labels extends DomElement {
      * @return the value of the label child.
      */
     @NotNull
-    @SubTag("label")
+    @SubTag(LABEL)
     GenericDomValue<String> getLabel();